requirements.txt
beautifulsoup4==4.3.2
Django==1.9.1
django-crispy-forms==1.6.0
django-cors-headers==1.1.0
django-registration-redux==1.4
requests==2.7.0
celery==3.1.18
redis==2.10.3
django-filter==0.13.0
Markdown==2.6.6
djangorestframework==3.3.3
django-oauth-toolkit==0.10.0